pyproject.toml
[tool.poetry]
name = "eventfor-us-backend"
version = "0.1.0"
authors = ["karuna <karuna@involve.software>"]
description = ""

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-debug-toolbar = "^4.3.0"
django-request-logging = "^0.7.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
environs = "^10.3.0"
factory-boy = "^3.3.0"
flake8 = "^7.0.0"
freezegun = "^1.4.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.0.0"
pytest-django = "^4.8.0"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
isort = "^5.10.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
